refactor: move classes related to SaaS migration under saas package
Renamed MigrationHandlerConfig into SaaSMigrationHandlerConfig
